Bug causes test run to come up as a non-debug run if you click on "Test" just as the project is loading.
This is a timing issue.  When project is openned, its automatically deployed, and the onDeploySuccess method will attempt to run it if the user clicks on run/test while its in the middle of deploying.  This scenario now gets the proper info on what button has been clicked to the onDeploySuccess handler which is separate from the standard handler used in project.js.  There is still redundant code here that should be cleaned up.


git-svn-id: https://src.springsource.com/svn/wm_svn/eng3/trunk@31630 bab4cc61-abf4-0310-9601-8b162d8b0fb1