Removes gap for scrollbar when no scrollbar needed
ISSUES: WM-4351

Dojo's grid reserves a bit of extra space for a scrollbar if it even suspects one may be needed.  So, after a short delay, tests to see if there was a scrollbar allocated, and if there was and clientHeight == scrollHeight, it gets rid of this by calling prerender().  There's a brief flicker, but better than hacking and maintaining the dojox.grid code.