Updates build and app initialization for phonegap Issues: WM-4523
The core change is to create a minimal build needed to initialize
the Application object, and then have a short 1ms delay to free up
the application initialization process (ie. the WebView component in IOS/Android), and then load remaining resources after the the WebView component has finished its thing.

1. Adds lib_build_phonegap and a wm_phonegap_misc build layers
2. Changes runtimeLoader and Application.js to load these
3. Changes application lifecycle to init components in doRun instead of postInit so that we can have processor idle time before we load most of the libraries
4. Updates the phonegap build to purge the lib_mobile files, and to add wm.isPhonegap to config.js