Now handles isDirty, save and getProgressIncrement calls Issues: WM-4497
Page wasn't yet fully incorporated into studio services tabs; did not update dirty indicator, did not have handling for closing of tabs.

NOTE: The current handling of dirty indicator is half-assed, it knows when a change has been made, but does not know when the change is unmade and we are back in our original state.