Fixes design time handling of composites on device change Issues: WM-4528
On device change, we regenerate widgets that may need to be rerendered for mobile design or desktop design.

Doing this without regenerating the Composite that owns it means that the any connect calls made by the Composite are lost.

This has been addressed by blocking the individual regeneration of components within the Composite (fix in Studio.js) and by flagging all Composites as needing to be completely recreated each time the device changes (which means all connects are reconnected).