Fixes type information for bind dialog Issues: WM-3934, WM-4536
Adds a typeFunc schema property so that if type can't be hard coded into the widget, it can be looked up at design time.

Updated DataSetEditor's dataValue and defaultInsert to use the typeFunc to get its current type.

Updated List dataSet type in the schema to be any "wm.Variable"