Insures user knows when their JSON is invalid Issues: WM-4574
I'm unhappy with how this fix tells user that their json is invalid
while they are in the middle of editing it, but at least that helps
them know why the save button is disabled.

The save button is disabled so that the user won't ever save an
invalid json string.