Refines how subobjects are client-side queried Issues: WM-4623
1. wm.Variable now will handle querying of subobjects
as long as those subobjects are not lists/arrays.

2. Grid and List setSelectedItem methods delete subobjects
so that comparisons are only on the main object for finding the
selected item.  Users are expected to pass in wm.Variables for
setSelectedItem, and we don't want to compare hundreds of related
objects that happen by chance to be part of the LiveView. Users
needing to query on related objects should use selectByQuery.